requirements.txt
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
alembic==1.13.2

celery==5.4.0
redis==5.0.8

selenium==4.23.1

google-genai==1.45.0

pytest==8.3.2
pytest-mock==3.15.1
httpx==0.28.1

python-dotenv==1.0.1
pydantic==2.11.7
pydantic-settings=2.11.0